requirements.txt
python-yaml
python-apt
python-setuptools
